+++ src/api/apps/vuelock/dto/create-vault.dto.ts
@@ -1,0 +1,6 @@
+import { IsNotEmpty } from "class-validator"
+
+export class CreateVaultDto {
+  @IsNotEmpty()
+  name: string
+}