+++ src/authorization/apikeyauthorizer/apikeyauthorizer.decorator.ts
@@ -6,6 +6,7 @@ import { EasenftTransactionModel } from "src/api/apps/easenft/entities/easenft.e
 import { FrostlakeAnalyticsModel } from "src/api/apps/frostlake/entities/frostlake-analytics.entity"
 import { SnowlakeTransactionModel } from "src/api/apps/snowlake/entities/snowlake.entity"
 import { SwapstreamTransactionModel } from "src/api/apps/swapstream/entities/swapstream.entity"
+import { VuelockSecretModel } from "src/api/apps/vuelock/entities/vuelock-secret.entity"
 import { WealthnowAssetModel } from "src/api/apps/wealthnow/entities/wealthnow-asset.entity"
 import { SubscriptionModel } from "src/api/subscription/entities/subscription.entity"
 import { apiPricing, subscriptionConfig } from "src/config/subscriptionConfig"
@@ -40,8 +41,18 @@ export const ApiKeyAuthorizer = createParamDecorator(
             const frostlakeUsedCredits = await FrostlakeAnalyticsModel.find({ apiKey }).countDocuments() * apiPricing.frostlake
             const snowlakeUsedCredits = await SnowlakeTransactionModel.find({ apiKey }).countDocuments() * apiPricing.snowlake
             const swapstreamUsedCredits = await SwapstreamTransactionModel.find({ apiKey }).countDocuments() * apiPricing.swapstream
+            const vuelockUsedCredits = await VuelockSecretModel.find({ apiKey }).countDocuments() * apiPricing.vuelock
             const wealthnowUsedCredits = await WealthnowAssetModel.find({ apiKey }).countDocuments() * apiPricing.wealthnow
-            const usedCredits = airlakeUsedCredits + cruxqlUsedCredits + dwalletUsedCredits + easeNftUsedCredits + frostlakeUsedCredits + snowlakeUsedCredits + swapstreamUsedCredits + wealthnowUsedCredits
+            const usedCredits = airlakeUsedCredits +
+              cruxqlUsedCredits +
+              dwalletUsedCredits +
+              easeNftUsedCredits +
+              frostlakeUsedCredits +
+              snowlakeUsedCredits +
+              swapstreamUsedCredits +
+              wealthnowUsedCredits +
+              vuelockUsedCredits
+
             const creditRequiredForCurrentRequest = apiPricing[`${requestedResource}`]
 
             if ((usedCredits + creditRequiredForCurrentRequest) < subscriptionConfig[`${subscription.selectedPlan.toLowerCase()}SubscriptionConfig`].grantedCredits) {