+++ src/api/apps/apps.module.ts
@@ -1,7 +1,6 @@
 import { Module } from "@nestjs/common"
 import { AirlakeModule } from "./airlake/airlake.module"
 import { FrostlakeModule } from "./frostlake/frostlake.module"
-import { WealthnowModule } from "./wealthnow/wealthnow.module"
 import { SwapstreamModule } from "./swapstream/swapstream.module"
 import { CruxqlModule } from "./cruxql/cruxql.module"
 import { DwalletModule } from "./dwallet/dwallet.module"
@@ -10,6 +9,6 @@ import { EasenftModule } from "./easenft/easenft.module"
 import { VuelockModule } from "./vuelock/vuelock.module"
 
 @Module({
-  imports: [AirlakeModule, FrostlakeModule, WealthnowModule, SwapstreamModule, CruxqlModule, DwalletModule, SnowlakeModule, EasenftModule, VuelockModule]
+  imports: [AirlakeModule, FrostlakeModule, SwapstreamModule, CruxqlModule, DwalletModule, SnowlakeModule, EasenftModule, VuelockModule]
 })
 export class AppsModule { }