+++ src/api/apps/wealthnow/dto/create-portfolio.dto.ts
@@ -1,6 +1,0 @@
-import { IsNotEmpty } from "class-validator"
-
-export class CreatePortfolioDto {
-  @IsNotEmpty()
-  name: string
-}