+++ src/config/envConfig.ts
@@ -12,7 +12,6 @@ export const envConfig = {
   frostlakeMongoDbUri: process.env.FROSTLAKE_DB_URI,
   snowlakeMongoDbUri: process.env.SNOWLAKE_DB_URI,
   swapstreamMongoDbUri: process.env.SWAPSTREAM_DB_URI,
-  wealthnowMongoDbUri: process.env.WEALTHNOW_DB_URI,
   vuelockMongoDbUri: process.env.VUELOCK_DB_URI,
   redisSocketHost: process.env.REDIS_SOCKET,
   redisPassword: process.env.REDIS_PASSWORD,